requirements.txt
pdfminer.six>=20221105
pytesseract
Pillow
pdf2image
notion-client
python-dotenv
requests
PyMuPDF
